+++ src/test/java/org/folio/circulation/api/requests/RequestsAPIStatusChangeTests.java
@@ -10,7 +10,6 @@ import org.folio.circulation.support.http.client.ResponseHandler;
 import org.hamcrest.core.Is;
 import org.junit.Test;
 
-import java.io.UnsupportedEncodingException;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.util.UUID;
@@ -28,8 +27,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID id = UUID.randomUUID();
 
@@ -54,8 +52,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID id = UUID.randomUUID();
 
@@ -80,8 +77,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID id = UUID.randomUUID();
 
@@ -106,8 +102,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID itemId = itemsFixture.basedUponSmallAngryPlanet(
       itemBuilder -> itemBuilder
@@ -170,8 +165,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID itemId = itemsFixture.basedUponSmallAngryPlanet(
       itemBuilder -> itemBuilder
@@ -235,8 +229,7 @@ public class RequestsAPIStatusChangeTests extends APITests {
     throws InterruptedException,
     ExecutionException,
     TimeoutException,
-    MalformedURLException,
-    UnsupportedEncodingException {
+    MalformedURLException {
 
     UUID itemId = itemsFixture.basedUponSmallAngryPlanet(
       itemBuilder -> itemBuilder